requirements (1).txt
gradio
langchain-google-genai
